% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignCluster.R
\name{AssignCluster}
\alias{AssignCluster}
\title{AssignCluster}
\usage{
AssignCluster(tab, cutoff = 0.5, return_in_cytotable = TRUE)
}
\arguments{
\item{g}{gene name}
}
\value{
.
}
\description{
assign clusters for each gene based on GO distance. By default, it incorporates this "group" into the cytoscape file
}
