% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff.R
\name{cutoff}
\alias{cutoff}
\title{cutoff}
\usage{
cutoff(data, method, stringency, Min, minparents)
}
\arguments{
\item{data}{.}

\item{method}{.}

\item{stringency}{.}

\item{Min}{.}

\item{minparents}{.}
}
\value{
value
}
\description{
.
}
