% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactTerms.R
\name{exactTerms}
\alias{exactTerms}
\title{Finds exact GO term annotation associated with a general description}
\usage{
exactTerms(M, term = "nuclear")
}
\arguments{
\item{M}{matrix of GO terms for each gene}

\item{term}{the general term that is being searched}
}
\value{
vector of exact terms that match the given term (ie: grep)
}
\description{
Akin to grep for the GO term annotations
}
