% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGenes.R
\name{findGenes}
\alias{findGenes}
\title{Find genes associated with a given GO term}
\usage{
findGenes(M, term = "nuclear")
}
\arguments{
\item{M}{Matrix of GO terms for each gene}

\item{term}{The GO term to call}
}
\value{
vector of gene names
}
\description{
Returns a vector of genes associated with a given GO term
}
