% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genes_in_cluster.R
\name{genes_in_cluster}
\alias{genes_in_cluster}
\title{genes_in_cluster
return a list of all genes assigned to a given cluster}
\usage{
genes_in_cluster(cluster = 1, tab = NULL)
}
\arguments{
\item{cluster, }{the cluster number}

\item{tab, }{cluster matrix}
}
\value{
list of genes belonging to a given cluster
}
\description{
genes_in_cluster
return a list of all genes assigned to a given cluster
}
