% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGo.R
\name{getGo}
\alias{getGo}
\title{Calculate the GO term matrix using hierarchical knowledge}
\usage{
getGo(genes, species = "mouse", preMinCol = 0, preMinRow = 0)
}
\arguments{
\item{genes}{genes to query GO terms for}

\item{species}{species to query from. Currently only supports human and mouse}

\item{preMinCol}{default = 0}

\item{preMinCol}{default = 0}
}
\value{
Matrix of GO terms per gene
}
\description{
.
}

