% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGO_bin.R
\name{getGo_bin}
\alias{getGo_bin}
\title{Calculate the GO term matrix without hierarchical information}
\usage{
getGo_bin(genes, species = "mouse", preMinCol = 0, preMinRow = 0)
}
\arguments{
\item{genes}{genes to query GO terms for}

\item{species}{species to query from. Currently only supports human and mouse}

\item{preMinCol}{default = 0}

\item{preMinCol}{default = 0}
}
\value{
Matrix of GO terms per gene
}
\description{
.
}

