\name{GONetwork-package}
\alias{GONetwork-package}
\alias{GONetwork}
\docType{package}
\title{
\packageTitle{GONetwork}
}
\description{
\packageDescription{GONetwork}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{GONetwork}
\packageIndices{GONetwork}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{GONetwork}

Maintainer: \packageMaintainer{GONetwork}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
