% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoTheDist.R
\name{GoTheDist}
\alias{GoTheDist}
\title{Calculate the distance matrix between genes based on GO annotation}
\usage{
GoTheDist(M, Min = 4, minparents = 8, MinRow = 0, Threads = NULL,
  method = "cosine")
}
\arguments{
\item{M}{matrix of GO terms for each gene}

\item{Min}{Cutoff to exclude low frequency GO terms. Minimum number of genes that contain the GO term. default = 4.}

\item{minparents}{Cutoff to exclude non-informative GO terms (ex: nucleus vs. Pol-II transcription factor). defualt = 8.}

\item{MinRow}{}
}
\value{
matrix in a cytoscape-friendly format
}
\description{
.
}

