% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InnerFun.R
\name{InnerFun}
\alias{InnerFun}
\title{Inner function for an internal parallelizable cosine distance calculation}
\usage{
InnerFun(j, x)
}
\arguments{
\item{j}{row in M2}

\item{x}{x-value in cosine}
}
\value{
cosine(x,y)
}
\description{
.
}
