% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_clusters.R
\name{loop_clusters}
\alias{loop_clusters}
\title{loop_clusters
loop through each cluster to test proportions}
\usage{
loop_clusters(CLUSTER, prop = 0.1, M = M, TAB = tab, return.max = 5)
}
\arguments{
\item{prop, }{proportion of genes required to share the term; default = 0.3}

\item{cluster, }{cluster id}
}
\value{
result of proportions test for all terms in a clusters
}
\description{
loop_clusters
loop through each cluster to test proportions
}
