% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OuterFun.R
\name{OuterFun}
\alias{OuterFun}
\title{Outer function for an internal parallelizable cosine distance calculation}
\usage{
OuterFun(i)
}
\arguments{
\item{j}{row in M2}

\item{x}{x-value in cosine}
}
\value{
cosine(x,y)
}
\description{
.
}
