% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbolConvert.R
\name{symbolConvert}
\alias{symbolConvert}
\title{symbolConvert}
\usage{
symbolConvert(x, in.type, out.type)
}
\arguments{
\item{x}{.}

\item{in.type}{.}

\item{out.type}{.}
}
\value{
value
}
\description{
.
}
