% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTerm.R
\docType{class}
\name{TermList-class}
\alias{TermList-class}
\title{topTerm
determine the most descriptive term for each cluster}
\arguments{
\item{tab, }{output from AssignCluster}

\item{M, }{matrix containing hierarchy information for all genes}
}
\value{
top term for each cluster assigned with AssignCluster
}
\description{
topTerm
determine the most descriptive term for each cluster
}
