% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_global_cluster_proportion.R
\name{test_global_cluster_proportion}
\alias{test_global_cluster_proportion}
\title{test_global_cluster_proportion
test the likelihood of a term belonging to a cluster}
\usage{
test_global_cluster_proportion(M, tab, t, cluster)
}
\arguments{
\item{t, }{term}
}
\value{
result of proportions test for a given term
}
\description{
test_global_cluster_proportion
test the likelihood of a term belonging to a cluster
}
