% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignCluster.R
\name{AssignCluster}
\alias{AssignCluster}
\title{AssignCluster}
\usage{
AssignCluster(tab, D, K = 10, rmv_stragglers = TRUE, cutoff = 0.3)
}
\arguments{
\item{tab, }{output from cyto()}

\item{K, }{number of groups to subset genes into}

\item{rmv_stragglers, }{run a second round of clustering after removing the unclustered genes default = TRUE}

\item{return_in_cytotable, }{returns the group numbers automatically into the original table provided from cyto(). Alternatively, you can set this to FALSE and simply return the group for each gene.}
}
\value{
.
}
\description{
assign clusters for each gene based on GO distance. By default, it incorporates this "group" into the cytoscape file
}
