% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignCluster.R
\name{AssignCluster}
\alias{AssignCluster}
\title{AssignCluster}
\usage{
AssignCluster(tab, cutoff = 0.1, return_in_cytotable = FALSE)
}
\arguments{
\item{tab, }{output from cyto()}

\item{cutoff, }{the minimum distance between nodes to be considered a cluster}

\item{return_in_cytotable, }{returns the group numbers automatically into the original table provided from cyto(). Alternatively, you can set this to FALSE and simply return the group for each gene.}
}
\value{
.
}
\description{
assign clusters for each gene based on GO distance. By default, it incorporates this "group" into the cytoscape file
}
