% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto.R
\name{cyto}
\alias{cyto}
\title{Convert distance matrix to a cytoscape formatted table}
\usage{
cyto(D, cutoff = 0.5)
}
\arguments{
\item{D}{matrix of gene distances based on GO terms}

\item{cutoff, }{minimum distance between nodes}
}
\value{
matrix in a cytoscape-friendly format
}
\description{
.
}

